feat(ai): 支持流式输出并实现长差异分块处理
- 支持流式响应 (Streaming)：重构了生成逻辑，支持通过回调函数实时接收并打印模型生成的文本，显著提升了用户交互体验。
- 引入分块处理机制：针对超过 8000 字符的长变更内容，采用“分块总结 -> 合并摘要 -> 生成最终消息”的策略，有效应对 LLM 上下文长度限制问题。
- 优化数据类型：将 `max_tokens` 从 `u16` 升级为 `u32`，以支持更大规模的 token 限制。
- 新增 `generate_summary` 功能：允许对代码变更的局部片段进行简短摘要提取，为分块处理提供支持。

+++ src/config.rs
@@ -15,7 +15,7 @@ pub struct Config {
     /// 当前选中的模型，默认使用 `model`
     #[serde(default = "default_selected_model")]
     pub selected_model: String,
-    pub max_tokens: u16,
+    pub max_tokens: u32,
     pub temperature: f32,
     pub auto_commit: bool,
     pub strict_format: bool,
